% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image_Analysis.R
\name{colonyThreshold}
\alias{colonyThreshold}
\title{Binary thresholding of gridded colony images}
\usage{
colonyThreshold(
  dir = getwd(),
  fileFormat = "jpg",
  files = list.files(path = dir, pattern = paste0("\\\\.", fileFormat), full.names = T),
  reference,
  outputFolder = T,
  plate.format = c(16, 24),
  verbose = "p",
  remove.noise = F,
  autorotate = F,
  inverse = T,
  contrast = NULL,
  grid.save = T,
  threshold.save = T,
  count.save = T,
  well.plate.format = c(8, 12)
)
}
\arguments{
\item{dir}{Directory of images. Defaults to current working directory.}

\item{fileFormat}{File format of the images. Defaults to \code{"jpg"}.}

\item{files}{Character vector of file names to be analysed. Defaults to all files of the specified \code{fileFormat} in the specified \code{dir}. Relative file paths to images in sub directories or full paths to images in other directories can also be specified.}

\item{reference}{File path of the image to be used as a grid reference.}

\item{outputFolder}{Logical. Should output files from the analysis be saved in a new directory. Defaults to \code{TRUE}.}

\item{plate.format}{Integer vector of length 2 specifying the number of rows and columns on the agar target plate, respectively. Defaults to \code{c(16, 24)}.}

\item{verbose}{Shows details about the results of running job. For detailed logs "l", for a progress bar "p" or for no output "n". Default is "l".}

\item{remove.noise}{Logical indicating noise/speckles should be remove from the thresholded image prior to analysis. Default is \code{FALSE}.}

\item{autorotate}{Logical indicating if image should be auto-rotated prior to processing. Only select this option if image is extremely rotated. gitter is able to handle small variations in rotations (1-2 degrees) without auto-rotating. Default is \code{FALSE}.}

\item{inverse}{Logical. Is the image colour inverted (i.e. colonies are darker compared to background). Defaults to \code{TRUE}.}

\item{contrast}{Integer between 1 and 100 indicating how much contrast should be applied to the image, prior to processing. A value of \code{NULL} will not apply any contrast. Default is \code{NULL}.}

\item{grid.save}{Logical. Should a gridded image be saved. Defaults to \code{TRUE}. This image is produced by \code{\link[gitter]{gitter}}.}

\item{threshold.save}{Logical. Should an image be saved showing which grid positions have been thesholded as colony present. Defaults to \code{TRUE}.}

\item{count.save}{Logical. Should an image be saved showing how many colonies are present at each dilution factor for each sample. Defaults to \code{TRUE}.}

\item{well.plate.format}{Integer vector of length 2 specifying the number of wells and columns of the liquid source plate, respectively. Dimensions must be factors of \code{plate.format} Defaults to \code{c(8, 12)}.}
}
\value{
\code{colonyThreshold} does not recturn any objects, but instead creates up to 4 files for each image analysed, as described below.
By default (\code{outputFolder=T}), files will be saved in a sub-directory called \code{Image_Analysis} within the current working directory.
\item{DAT file}{
 A tab delimited file containing quantified colony sizes, as described in \code{\link[gitter]{gitter}}.
 An sixth column marking whether a colony has been classified as present (1) or absent (0) is added.
}
\item{Gridded image}{
 As described in \code{\link[gitter]{gitter}}.
}
\item{Threshold image}{
 Thresholded image showing whether a colony has been classified as present or absent for each position on the plate.
}
\item{Count image}{
 Image showing how many colonies have been classified as present for each sample at each dilution factor.
}
}
\description{
\code{colonyThreshold} analyses a batch of images of gridded colonies and quantifies whether or not there is a colony in each position. Grid identification and colony size quantification is done using the \code{\link[gitter]{gitter.batch}} function, which is then followed by additional thresholding to classify whether a colony is present or absent at each position.
}
\examples{
#Get the directory of files to be analysed
dir <- system.file("extdata", "images", package="DeadOrAlive")

#Get the reference image
reference <- system.file("extdata", "reference.jpg", package="DeadOrAlive")

#Analyse the files to identify whether there is a colony or not in each position
#Note: This will create a new directory called 'Image_Analysis'
\dontrun{
colonyThreshold(dir=dir, reference=reference)
}
}
