% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Rearrangement.R
\name{extractColonyVectors}
\alias{extractColonyVectors}
\title{Extraction of colony patterns from analysed images}
\usage{
extractColonyVectors(
  dir = getwd(),
  plateReferenceFile,
  sampleReferenceFile,
  plate.format = c(16, 24),
  well.plate.format = c(8, 12),
  byRow = T
)
}
\arguments{
\item{dir}{Directory of DAT files. Defaults to current working directory.}

\item{plateReferenceFile}{Path to a csv file containing information on the identity of each plate. This file must have 3 columns as follows:
\itemize{
\item{\code{time}: Numerical values indicating the time in days.}
\item{\code{plate}: Characters indicating the name of each plate.}
\item{\code{fileName}: Full or relative file paths of each DAT file to analyse. Only DAT files listed here will be analysed.}
}}

\item{sampleReferenceFile}{Path to a csv file containing information on the samples present on each plate. Each column must correspond to a different plate, with the column names corresponding to the plates listed in the \code{plate} column of the \code{plateReferenceFile}.
Each sample name should be unique. The number of samples in each column should correspond to the number of samples on each plate.
In the case that \code{byRow} is \code{TRUE}, this should mean that there is one sample for each column on the plate. For example, if colonies were pinned from a 96-well plate, this would mean 8 samples.
Conversely, in the case that \code{byRow} is \code{FALSE}, this should mean that there is one sample for each row on the plate. For example, if colonies were pinned from a 96-well plate, this would mean 12 samples.}

\item{plate.format}{Integer vector of length 2 specifying the number of rows and columns on the agar target plate, respectively. Defaults to \code{c(16, 24)}.}

\item{well.plate.format}{Integer vector of length 2 specifying the number of wells and columns of the liquid source plate, respectively. Dimensions must be factors of \code{plate.format} Defaults to \code{c(8, 12)}.}

\item{byRow}{Do dilution series run across the rows of the plate (i.e left to right)? Defaults to \code{TRUE}. Set to \code{FALSE} if dilution series run across columns (i.e. top to bottom).}
}
\value{
An object of class \code{colonyVectors}, containing the patterns extracted from each plate.
}
\description{
\code{extractColonyVectors} analyses a batch of DAT files produced by \code{\link{colonyThreshold}} in order to extract colony patterns for each sample.
}
\examples{
#Get a csv file showing the identity of each plate to be analysed
plateReferenceFile <- system.file("extdata", "plateReferenceFile.csv", package="DeadOrAlive")

#Get a csv file showing the identity of each sample on each plate
sampleReferenceFile <- system.file("extdata", "sampleReferenceFile.csv", package="DeadOrAlive")

#Get the directory of files to be analysed
dir <- system.file("extdata", "Image_Analysis", package="DeadOrAlive")

#Get the patterns of colonies from the files (from left to right across the plate)
myColonyVectors <- extractColonyVectors(dir=dir, plateReferenceFile, sampleReferenceFile)
}
